microservicios/edits/requirements_updated.txt
# PUCP Cloud Orchestrator - Requirements (UPDATED with OpenStack Service)

# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0

# Linux cluster management
libvirt-python==9.0.0

# OpenStack clients (para el microservicio)
keystoneauth1==5.2.0
python-novaclient==18.3.0
python-neutronclient==9.0.0
python-glanceclient==4.3.0
python-cinderclient==9.3.0

# WSGI server
gunicorn==21.2.0

# Python standard library (no need to install but listed for reference)
# sqlite3
# uuid
# ipaddress
# datetime
# typing
# json
# os
# sys
# logging
# threading
# time
# subprocess

# Optional: Monitoring and logging
# structlog==23.1.0
# prometheus-client==0.17.1

# Optional: Development and testing
# pytest==7.4.0
# pytest-flask==1.2.0
# pytest-cov==4.1.0